Added auth.is_logged_in() function and ensure  always signaled
New function reduces need to look at auth internals, as well as supports better handling for when requests fail or don't complete. With this change, a developer can always depend on the signal  being raised, so they can handle the callback state as needed. This will indeed emit this signal on more code branches than before, which should be properly tested for any unexpected behavior for anyone who depends on connecting a response function to  and was already somehow separpately handling these branch cases